Added: try to open email attachments that have no file extension:
Instead of popping up an error saying unknown extension, try to
open the file with the externel viewer. Let the external
viewer decide if the file can be opened or not.

This has not been super heavily tested, although I was able to
open various plain text attachments without issues, which had
no extension, like "mailcap", "config" and so on.

(I used w3m as my externel viewer in tests, following the fix in
a previous commit to make it work OK.)